iccid_imei_export/requirements.txt
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-api-python-client>=2.0.0

